requirements.txt
# Python dependencies to be installed using pip http://www.pip-installer.org/

# Web server.
gunicorn==18.0
gevent==0.13.8

# App Framework
pyramid==1.4
pyramid_hsts==1.1
pyramid_redis==0.1
pyramid_tm==0.7

# Database.
psycopg2==2.5.1
gevent-psycopg2==0.0.3
sqlalchemy==0.8.2
zope.sqlalchemy==0.7.3
alembic==0.6

# Redis client
redis==2.8.0

# Testing
coverage==3.6
mock==1.0.1
nose==1.3.0
WebTest==2.0.7

# Utils.
bpython==0.12
requests==1.2.0

# Install this egg.
-e .
